python_producer/requirements.txt
confluent-kafka
python-dateutil
faker
kafka-python==2.0.2